Hacking PvP replay for war attacks

+++ src/main/java/swcnoops/server/model/DeploymentData.java
@@ -3,9 +3,9 @@ package swcnoops.server.model;
 import java.util.Map;
 
 public class DeploymentData {
-    public Map<String, Long> troop;
-    public Map<String, Long> specialAttack;
-    public Map<String, Long> hero;
-    public Map<String, Long> champion;
+    public Map<String, Integer> troop;
+    public Map<String, Integer> specialAttack;
+    public Map<String, Integer> hero;
+    public Map<String, Integer> champion;
     public Map<String, String> creature;
 }
